migration/requirements.txt
# https://pypi.org/project/pylodstorage/
pylodstorage>=0.0.29
# https://pypi.org/project/python-dateutil/
python-dateutil
# https://pypi.org/project/py-3rdparty-mediawiki/
py-3rdparty-mediawiki>=0.4.8
# https://github.com/5j9/wikitextparser
wikitextparser>=0.47.04